# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UNZA DigiHealth API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            $ref: "#/components/schemas/UserRole"
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    put:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/status:
    patch:
      operationId: updateUserStatus
      tags: [users]
      summary: Update user status (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserStatusRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Consultations
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserRole:
      type: string
      enum: [student, doctor, pharmacist, lab_technician, mental_health_counselor, hiv_professional, admin]
    UserStatus:
      type: string
      enum: [active, inactive, suspended]
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        studentNumber:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        status:
          $ref: "#/components/schemas/UserStatus"
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, status, createdAt]
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        studentNumber:
          type: string

    # Consultation
...
    UpdateUserStatusRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/UserStatus"
      required: [status]
...